feat: allow --debug anywhere on the command line
